task1/task1_requirements.txt
paho-mqtt~=1.6.1
psycopg2~=2.9.3